Move Storyblok components to Storyblok folder
